Change 'const float* input' to 'const int8_t* input'

+++ src/operators/fully-connected-nc.c
@@ -3136,7 +3136,7 @@ enum xnn_status xnn_setup_fully_connected_nc_qdu8_f16_qc8w(
 }
 
 enum xnn_status xnn_setup_fully_connected_nc_qp8_f32_qc4w(
-    xnn_operator_t fully_connected_op, const float* input, float* output,
+    xnn_operator_t fully_connected_op, const int8_t* input, float* output,
     void* workspace) {
   return setup_fully_connected_nc(
       fully_connected_op, xnn_operator_type_fully_connected_nc_qp8_f32_qc4w,
@@ -3144,7 +3144,7 @@ enum xnn_status xnn_setup_fully_connected_nc_qp8_f32_qc4w(
 }
 
 enum xnn_status xnn_setup_fully_connected_nc_qp8_f32_qc8w(
-    xnn_operator_t fully_connected_op, const float* input, float* output,
+    xnn_operator_t fully_connected_op, const int8_t* input, float* output,
     void* workspace) {
   return setup_fully_connected_nc(
       fully_connected_op, xnn_operator_type_fully_connected_nc_qp8_f32_qc8w,
@@ -3152,7 +3152,7 @@ enum xnn_status xnn_setup_fully_connected_nc_qp8_f32_qc8w(
 }
 
 enum xnn_status xnn_setup_fully_connected_nc_qp8_f32_qb4w(
-    xnn_operator_t fully_connected_op, const float* input, float* output,
+    xnn_operator_t fully_connected_op, const int8_t* input, float* output,
     void* workspace) {
   return setup_fully_connected_nc(
       fully_connected_op, xnn_operator_type_fully_connected_nc_qp8_f32_qb4w,
